delete dialog + bug fixing

+++ lib/utils/remindersHelper.dart
@@ -16,19 +16,16 @@ class RemindersHelpers {
 
   Future deleteReminder(BuildContext context, String docID) async {
     try {
-      bool delete = await deleteDialog(context);
-      if (delete) {
-        String userID = await Profile().getUserID();
-        DocumentSnapshot snapshot =
-            await users.doc(userID).collection("reminders").doc(docID).get();
-        if (snapshot.exists) {
-          await users.doc(userID).collection("reminders").doc(docID).delete();
-          if (context.mounted) {
-            dialog(
-              context,
-              "Delete Successful",
-            );
-          }
+      String userID = await Profile().getUserID();
+      DocumentSnapshot snapshot =
+          await users.doc(userID).collection("reminders").doc(docID).get();
+      if (snapshot.exists) {
+        await users.doc(userID).collection("reminders").doc(docID).delete();
+        if (context.mounted) {
+          dialog(
+            context,
+            "Delete Successful",
+          );
         } else {
           if (context.mounted) {
             dialog(
@@ -215,14 +212,16 @@ class RemindersHelpers {
                 direction: reminder['userID'] == reminder['currentID']
                     ? DismissDirection.endToStart
                     : DismissDirection.none,
-                onDismissed: (direction) {
-                  RemindersHelpers().deleteReminder(
-                    context,
-                    reminder['reminderID'],
-                  );
-                  if (context.mounted) {
-                    GoRouter.of(context).pushReplacement("/reminders");
-                  }
+                confirmDismiss: (DismissDirection direction) {
+                  return confirmDismissDialog(context, () {
+                    if (context.mounted) {
+                      RemindersHelpers().deleteReminder(
+                        context,
+                        reminder['reminderID'],
+                      );
+                      GoRouter.of(context).pushReplacement("/reminders");
+                    }
+                  });
                 },
                 background: Container(
                   color: Colors.red,
